Rename cell to description
