+++ src/ipss_cell_description.hpp
